% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zimed.R
\name{zimed}
\alias{zimed}
\title{The effect decomposition of mediation analysis for zero-inflated models in microbiome data}
\usage{
zimed(M_mat, Treat, Outcome, method, ci.method)
}
\arguments{
\item{M_mat}{a data frame or matrix of the microbiome count data. Rows of the matrix represent samples and columns are the taxa}

\item{Treat}{a binary vector of group indicators}

\item{Outcome}{a continuous outcome}

\item{method}{methods for testing the H0 hypothesis, including joint, HDMT or DACT}

\item{ci.method}{methods for calculating confidence interval, including bootstrap or delta}
}
\value{
NIE the estimated natural indirect effect

NIE.p the calculated p value for NIE

NIE.ci the calculated confidence interval for NIE

NIEA the estimated natural indirect effect through changes of abundance

NIEA.p the calculated p value for NIEA

NIEA.ci the calculated confidence interval for NIEA

NIEP the estimated natural indirect effect through changes of absence/presence

NIEP.p the calculated p value for NIEP

NIEP.ci the calculated confidence interval for NIEP
}
\description{
The effect decomposition of mediation analysis for zero-inflated models in microbiome data
}
